pyproject.toml
[tool.poetry]
name = "biosample-xmldb-sqldb"
version = "0.1.0"
description = ""
authors = ["Mark A. Miller <MAM@lbl.gov>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
pandas = "^2.2.0"
sqlalchemy = "^2.0.26"
pyarrow = "^15.0.0"
psycopg2-binary = "^2.9.9"
biopython = "^1.83"
lxml = "^5.1.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
split-into-N-biosamples = "biosample_xmldb_sqldb.split_into_N_biosamples:cli"
#pivot-harmonized-attributes = "pivot_harmonized_attributes.split_into_N_biosamples"
pivot-from-bp-id = "biosample_xmldb_sqldb.pivot_from_bp_id:main"